requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.0
django-otp==1.5.4
packaging==24.2
pipdeptree==2.24.0
pydantic_core==2.27.2
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
